Move PackageWrapper and associated functions to its own package
