+++ pkg/pkg/testdata/getOverlayFiles/overlay/dir1/file3.txt
